(#1) Develop display layout
