+++ shinoa/lib/view/keyboard.dart
@@ -1,0 +1,21 @@
+
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class Keyboard extends StatefulWidget {
+  const Keyboard({Key? key}) : super(key: key);
+
+  @override
+  _KeyboardState createState() => _KeyboardState();
+}
+
+class _KeyboardState extends State<Keyboard> {
+  @override
+  Widget build(BuildContext context) {
+    return Expanded(
+      child: Container(
+        color: Colors.red,
+      ),
+    );
+  }
+}
